Lunch Update 4/16
Finished rso_details, join and leave abilities for rso, index and welcome pages.

Created session.php and config.php, which hold session data, such as $u_id and all of the database information accessed by every page. Session.php also ensures that a user is logged in, redirecting them to index.php if they are not.